Add gamma correction to the LED API
This means that real-life images, e.g. from the camera, will look much better - human sight is nonlinear, and gamma curves account for this.

Note: it may be slightly better (more colour resolution would be retained) to pack into the framebuffer using the current linear colour space and then perform the gamma conversion when writing to the LEDs.