Added method to clear all outstanding events.
Clearing the outstanding events is useful if you need
to get rid of a button release event in preparation for receiving
the next button press event. Since the release event happens after
the previous call to wait_for_event() or get_events() it isnt cleared
and will appear in the next call to wait_for_event() or get_events().